pyproject.toml
[project]
name = "task-rider"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "PyQt6",
    "darkdetect",
    "attrs",
    "cattrs",
    "pyremindkit@git+https://github.com/namuan/pyremindkit",
    "schedule",
    "pyobjc",
]

[dependency-groups]
dev = [
    "black",
    "flake8",
    "pyinstaller>=6.17.0",
    "ruff>=0.14.10",
]

[project.scripts]
task-rider = "app.__main__:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.uv]
package = true

[tool.setuptools]
packages = ["app"]
